refactor(user_externalid): move method to groups module
Move method to semantically similar group `groups`.
A seperate file for this method is not necessary.
