+++ invenio_rdm_pure/source/rdm/run/groups.py
@@ -10,6 +10,9 @@
 import json
 import os
 
+from flask_login import current_user
+from invenio_oauthclient.models import UserIdentity
+
 from ...general_functions_source import add_spaces
 from ...pure.requests_pure import get_pure_metadata
 from ...reports import Reports
@@ -17,6 +20,16 @@ from ..database import RdmDatabase
 from ..general_functions import GeneralFunctions
 from ..record_manager import RecordManager
 from ..requests_rdm import Requests
+
+
+def user_externalid():
+    """Description."""
+    if current_user.is_authenticated:
+        id = current_user.get_id()
+        user_external = UserIdentity.query.filter_by(id_user=id).first()
+        if user_external:
+            return user_external.id
+    return False
 
 
 class RdmGroups:
